requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
bitarray==3.8.0
certifi==2026.2.25
charset-normalizer==3.4.4
ckzg==1.0.2
cytoolz==1.1.0
dataclasses-json==0.6.7
dateparser==1.3.0
distro==1.9.0
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hexbytes==0.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.16
langchain-openai==1.1.10
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.7
lru-dict==1.2.0
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.2
openai==2.24.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
parsimonious==0.10.0
pillow==12.1.1
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
poly_eip712_structs==0.0.1
propcache==0.4.1
protobuf==7.34.0
py_builder_signing_sdk==0.0.2
py_clob_client==0.34.6
py_order_utils==0.3.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python_binance==1.0.35
pytz==2026.1.post1
pyunormalize==17.0.0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rlp==4.1.0
rpds-py==0.30.0
simsimd==6.5.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.47
tenacity==9.1.4
tiktoken==0.12.0
toolz==1.1.0
tqdm==4.67.3
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
web3==6.14.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
